--- media/img/logo_uHVcH5Y.jpg
@@ -1,1 +1,0 @@
-content
