--- media/img/logo_uMNuOTR.jpg
@@ -1,1 +1,0 @@
-content
