--- media/img/logo_vq2KzQC.jpg
@@ -1,1 +1,0 @@
-content
