--- media/img/logo_vxnR2Vs.jpg
@@ -1,1 +1,0 @@
-content
